requirements.txt
Django>=4.2.0
firebase-admin>=6.0.0
pyrebase4>=4.6.0
requests>=2.25.0
